refactor(dash): split runner and measurer
This diff splits dash.go into runner.go and measurer.go.

By splitting by topic, we get easier to test and to modify code.

This PR is yak shaving for https://github.com/ooni/probe/issues/2398.
